% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{la_fit_mirror}
\alias{la_fit_mirror}
\alias{la_fit_nonmirror}
\title{Fit "Mirror" and "Non-Mirror" models (wrappers for readability)}
\usage{
la_fit_mirror(x, y, engine = "glm", ...)

la_fit_nonmirror(x, y, engine = "glm", ...)
}
\arguments{
\item{x}{data.frame/matrix of features}

\item{y}{outcome vector}

\item{engine}{character: "glm" or "glmnet"}

\item{...}{additional arguments passed to la_fit}
}
\value{
langAssessR_model object
}
\description{
Fit "Mirror" and "Non-Mirror" models (wrappers for readability)
}
