% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{la_predict}
\alias{la_predict}
\title{Predict with a langAssessR model}
\usage{
la_predict(model, newx, type = NULL, threshold = 0.5)
}
\arguments{
\item{model}{object from la_fit()}

\item{newx}{data.frame/matrix with same feature columns as training (order flexible)}

\item{type}{"prob" (classification default) or "response" for regression; "label" returns 0/1 labels at threshold}

\item{threshold}{classification threshold for "label" predictions (default 0.5)}
}
\value{
numeric vector (probabilities or predictions), or integer(0/1) if type="label"
}
\description{
Predict with a langAssessR model
}
