% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_attention_heatmap}
\alias{plot_attention_heatmap}
\title{Plot attention heatmap}
\usage{
plot_attention_heatmap(df)
}
\arguments{
\item{df}{data frame with start and weight columns}
}
\description{
Plot attention heatmap
}
