% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_auc_by_group}
\alias{plot_auc_by_group}
\title{Plot AUC by group}
\usage{
plot_auc_by_group(df)
}
\arguments{
\item{df}{data frame with subgroup, auc, lcl, ucl columns}
}
\description{
Plot AUC by group
}
