% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_calibration_by_group}
\alias{plot_calibration_by_group}
\title{Plot calibration by group}
\usage{
plot_calibration_by_group(df)
}
\arguments{
\item{df}{data frame with mean_pred, obs_rate, subgroup columns}
}
\description{
Plot calibration by group
}
