% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_decision_curve}
\alias{plot_decision_curve}
\title{Plot decision curve}
\usage{
plot_decision_curve(df)
}
\arguments{
\item{df}{data frame with threshold, net_benefit, policy columns}
}
\description{
Plot decision curve
}
