% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_deid_counts}
\alias{plot_deid_counts}
\title{Plot de-identification counts}
\usage{
plot_deid_counts(df)
}
\arguments{
\item{df}{data frame with entity and count columns}
}
\description{
Plot de-identification counts
}
