% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_error_balance}
\alias{plot_error_balance}
\title{Plot error balance}
\usage{
plot_error_balance(df)
}
\arguments{
\item{df}{data frame with subgroup, fnr, fpr columns}
}
\description{
Plot error balance
}
