% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_forest_compare}
\alias{plot_forest_compare}
\title{Plot forest comparison}
\usage{
plot_forest_compare(df, metric_label = "AUROC")
}
\arguments{
\item{df}{data frame with site, estimate, model, lcl, ucl columns}

\item{metric_label}{label for x-axis}
}
\description{
Plot forest comparison
}
