% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_site_performance}
\alias{plot_site_performance}
\title{Plot site-wise performance}
\usage{
plot_site_performance(df, metric_label = "AUROC")
}
\arguments{
\item{df}{data frame with site, estimate, lcl, ucl columns}

\item{metric_label}{label for y-axis}
}
\description{
Plot site-wise performance
}
