% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_color_lang}
\alias{scale_color_lang}
\alias{scale_fill_lang}
\title{Color scales for langAssessR}
\usage{
scale_color_lang(...)

scale_fill_lang(...)
}
\arguments{
\item{...}{arguments passed to scale_color_manual/scale_fill_manual}
}
\description{
Color scales for langAssessR
}
