% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_lang}
\alias{theme_lang}
\title{Global theme for langAssessR figures}
\usage{
theme_lang(base_size = 11, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Global theme for langAssessR figures
}
